requirements.txt
cairosvg==2.7.1
Pillow==10.2.0 